+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Expected: May 2026
 ### Features
 
 * New: XPath translate()
+* New: [Default values for YANG leaf-list](https://github.com/clicon/clixon/issues/664)
 * Enumerated types now appear in CLI help texts, see eg https://github.com/clicon/clixon/issues/183
 * show memory: Added detailed statistics for config datastores (CLI and RPC)
 * New `clixon-config@2026-03-01.yang` revision